chore: reorganise and document scpi code
